Add admin and user checks to GET /orders
